+++ src/arviz_plots/plots/__init__.py
@@ -1,6 +1,6 @@
 """Batteries-included ArviZ plots."""
 
-from .posteriorplot import plot_dist
+from .distplot import plot_dist
 from .traceplot import plot_trace
 
 __all__ = ["plot_dist", "plot_trace"]
