+++ src/arviz_plots/plots/distplot.py
